Implemented: Drag to area for type reassignment
close #1

Canvas is now divided into colored areas. There is one region for each of the following Decision Element types:
- Action
- Intermediate
- Outcome
- External

Area bounds are defined as percentages of the cavnas width/height, so that canvas dimensions can be changed.
When a node is moved into a region, it inherits the categorization of that region. This functionality is attached to the "change:position" event listener for all nodes.

Important note: On load, diagrams elements are not made to fit within the designated Decision Element regions. Nor are they updated based on the regions they load into.
Saving the diagram also does not validate or update element types based on region.
Element types for each node are only updated when that node is moved.

---

Other notes:

- All functionality for Decision Element Areas is in decisionElementArea.js
- Moved functionality related to saving JSON data from index.js to fileIO.js
- Updated cdd.json to align with the defined Decision Element Area bounds