Improved saving
FileIO saving function no longer assumes it is working with the same set of elements as was in the initial JSON.
Now original JSON is saved in a map from UUID to JSON Object.
Then, when saving the current diagram, that map is checked for elements that already existed. If found, it selectively overwrites properties in the existing JSON.
If not found, it constructs a new JSON representation for the new element.
This way, any potential addon information for existing elements is preserved.