Implemented element selection and deselection
Implemented a SelectionBuffer class that tracks the last N selected diagram elements.
Decision Elements now define two sets of visual attributes, for selected/unselected state.
Selection buffer will be useful for deleting elements and creating/deleting dependencies later.

Selection buffer is resizeable for multi-selection. Currently there's no mechanism for activating multi-selection,
but increasing buffer size to 2 or more does work as expected!

User can select an item (or more if buffer size is increased) by clicking an item.
This is done on mouse-up, because the click event was too sensitive. Many quick
sequences of clicks would be too motion-heavy to activate the click event. Because
of this, dragging an element also selects it when the drag is complete. I don't really
mind this, so I left it.

User can de-select all selected items by clicking a blank spot on the canvas.
(Also done w/ mouse-up, for the same reasons as above)