Implemented dependecy toggle Added a UI button that toggles dependencies between elements. Multi-select 2 elements in order, source -> target, then click the button to toggle a dependency between those elements.
For 3 or more multi-selected elements, this button creates/removes dependencies in an A -> B -> C -> D fashion,
where elements were selected in alphabetic order.
Toggle prefers to add dependencies if possible, so if any dependencies are missing along the chain they will be added.
In other words, toggle only deletes dependencies if a chain already fully contains the dependencies implied by the selection.