Changes to JSON Editor view apply to JointJS graph
Changes made in the JSON Editor view are now mirrored in the JointJS graphical view.

Upcoming:
The reverse is not yet true (changes made to the graph do not update the JSON Editor view).
No JSON Schema validation is performed yet, so the user can supply bad values to the JSON Editor.