Changes to JointJS graph apply to JSON Editor view
Changes made in the JointJS graphical view are now mirrored in the JSON Editor view.

package.json:
svelte-jsoneditor requires that JSON initially passed to it not be mutated.
So, added lodash-es package for cloneDeep to make mutable changes to a copy, then pass it in.

fileIO.js:
Updated functions to no longer be specialized for the Download button.
Now able to get new JSON for latest CDD on demand. index.js uses this to send changes to JSON Editor.

index.js:
- Refactor: runtimeData now used consistently in its initial object form throughout.
- Refactor: Ensure json editor's JSON object does not get mutated throughout.
- Added missing JSDoc documentation

TODO:
- Schema validation for JSON Editor
- Selecting elements on graph affects expand/collapse groups on JSON Editor (for "highlighting")

MAYBE:
- Refactor: Always pass JSON Editor a cloneDeep copy of the JSON data as late as possible (just before editor.update or editor initialization)
    - This would allow mutative changes to that JSON data everywhere else. i.e. give the editor a copy, not the original.