Selections in JSON Editor carry over to Graph
When user selects elements in the JSON Editor view, these elements are selected/highlighted in the graph view.
This further eases navigation between graph and JSON data,
allowing user to easily see which portions of the graph come from which portions of the JSON data.

User can select any sub-element of a Decision Element. Graph will figure out what element the sub-element belongs to and highlight that element.
If a user selects a range of elements in the json view, all those elements will be selected in the graph view.

Also:
- Fixed an issue with graph selection -> json selection mirroring. Disabled this behavior in non-tree view modes, where expand and collapse functions are unsupported.
- Specified "normal" router and connector for dependency links. This doesn't change behavior but I want to highlight these options for myself in the source in case they need changing later.