Containerization proof of concept
Implemented a minimal proof-of-concept for use with [model-api-test](https://github.com/opendi-org/model-api-test) (repo currently private).
Build Docker images with the given dockerfiles, then run both containers. Navigate web browser to http://localhost:3000 to access the CDD web frontend locally.
This frontend demonstrates connection with the API project by attempting to fetch the Coffee CDD from the API.

If the frontend is successful, you should see the copy of the CDD from the API database. If unsuccessful, frontend falls back on loading from the local JSON file.
Frontend attempts to fetch the model directly by the model's UUID, so if the model is not first added to the local database, this will always fail.
Add the model to the database by sending a POST request with the coffee CDD (or any other CDD) in the payload. Recommend modifying the CDD title so that it can be differentiated from the local copy in the JSON file.

dockerfile facilitates a 2-stage build process.
First, creates a temporary NPM image and container, with all project files copied over. This container builds the project.
Next, creates the final runtime container. Copies the build artifacts from the temporary container to the final one.
Final container is an nginx webserver, with entry point set to serve the build artifacts on localhost:80.
Recommend binding port 80 on the container to port 3000 on the host machine. Containerized webapp will assume the API is accessible on localhost port 8080.