nginx -> Apache web server. Started API class.
Adjusted dockerfile to use an Apache web server.

Started an API class to perform raw API stuff and return well-formed JSON / etc.
Currently can only fetch models by UUID. Need to work on adding more endpoints to this.

Adjusted index.js for proof of concept API interaction. Currently just tries to fetch a specific model.
If this model is pushed to the API, it will fetch properly. If fetch fails, it will fall back on the old packed-in JSON file.