React refactor: Convert validation to TypeScript
Took existing validation logic in `validation.js` and converted it to TypeScript in `validation.tsx`.
Re-implemented it by adding the new TypeScript-friendly validator to the `validator` prop on the React wrapper for svelte-jsoneditor.
Also re-added the additional logic needed to only update the common source of truth for model JSON if the updated JSON from the editor passes AJV validation and my additional UUID checks. This went in the existing onChange prop.

These changes required these packages be added:
- ajv
- ajv-formats
- uuid