React refactor: Add basic select/deselect
Reimplemented the idea of a selection buffer.
Selection Buffer state is managed by `CausalDecisionDiagram.tsx`. It is a simple list of strings, containing UUIDs of selected elements.

Logic for updating the selection buffer with select/deselect/clear actions lives in `src/lib/updateElementSelection`.

Each element handles updating its own selection state in the buffer, via new hooks for accessing the selectionBuffer state
and for selecting an element.
Elements now have a checkbox in the top-right for selection. Checkbox includes a label for selection order. This
should make it easier to keep track of dependency arrow direction when toggling is added back in later.

Moved some of the styling for decision elements to `index.css`.
Defined a new `selected-diagram-element` style class that is applied when an element's selection state is `true`.
Turns the border and top-bar (dragging handle) white.

CausalDecisionDiagram handles styling for dependency arrows, so it now checks the selection buffer to see if an
arrow is connected to a selected element, then turns it white.
Added a new "causal type color" for selected arrows specifically. Highlight color for borders and draggable handle had to go in `index.css`, so these are split :(

Slightly adjusted overall layout for CausalDecisionDiagram. Added a new `div` that sits under the diagram contents, over the background of the diagram-containing `div`.
This new `div` catches `onClick` commands that are specific to the background. Its position as a leaf in the DOM tree means that no other click events will propagate up to it.
This prevents click events from propagating up from diagram elements and being handled as a background click, clearing the selection buffer.