Add model meta info box, top-left of diagram area
Added an info box in the style of the floating controls legend, for model meta info.
Currently shows model name, diagram name, and summary.
If model or diagram are not named, uses placeholder label.
If no summary exists, does not show the option to view the summary.

Summary view is collapsible.

Other changes:
- Added more controls to the controls legend
- Moved some CDD elements in coffee.json so they're not hidden behind the new meta info box
- Capitalized coffee decision model name. Slight reword in summary.
- Rounded the bottom corners of all CDD elements. I think it looks neat.