React refactor: Mirror selections: graph -> JSON
Reimplemented selection mirroring between the graphical view and the JSON editor view.

`lib/getExpandedPathsForSelectedElements.tsx`:
Implements the logic for getting the JSON paths of everything that needs to be expanded in the JSON editor.
The default expanded paths are also specified here.

`VanillaJSONEditor.tsx`:
Added new optional prop `expandedPaths`.
This is expected to be a useState. When changed, a useEffect updates the expanded/collapsed paths for the editor.
By default, editor collapses all paths except for those specified by `expandedPaths`.
Editor scrolls to the last path in `expandedPaths`.

`CausalDecisionDiagram.tsx`:
Added new prop `setExpandedPaths`.
This is expected to be the set function for a useState. A useEffect dependant on `selectionBuffer` updates
the JSON editor's expanded paths when graphical selection changes.
Added a new useMemo that maps element UUIDs to a list of their associated dependencies.
Associated dependencies for a given element have that element as either a source or a target.

`App.tsx`:
Initialize the new useState for expanded paths, pass it to CDM and JSON Editor.

`EditorAndHelpMenu.tsx`:
Pass `expandedPaths` on to `VanillaJSONEditor`.