New element (connected) implemented
Implemented:
When one or more elements are selected, and the New Element button is clicked, the new
element created will be a dependency TARGET of all the selected elements.

The new element becomes the only thing in the selection buffer at that point.

So if you want to create a chain quickly, you can just click the New Element button a lot.
If you want to create a set of unconnected elements, you have to click the background in between
each press of New Element.

`elementCRUD.tsx`:
- Add new functionality to `addNewElement` for if the `selectionBuffer` has items in it
- `addNewElement` now needs access to the diagram element map, to get element names

`ElementCRUDPanel.tsx`:
- Pass diagram elements map along to `addNewElement`

`CausalDecisionDiagram.tsx`:
- Pass diagram elements map along to `ElementCRUDPanel.tsx`