Fix: 0th dependency never expands in JSON editor
Here was the issue:
When selecting an element that is either a source or target in the dependency at index 0 in model.diagrams.dependencies,
the JSON for that dependency is never expanded. It is supposed to be expanded in these cases.

It was not getting added to the list of paths that the JSON editor is meant to expand.

This was a "0 evaluates to false" error. Whoops!
Replaced
if(depIdx)
with
if(depIdx !== undefined)

where depIdx is populated by a function that can return undefined (if it finds no index for the dependency).