Style: Make diagram elements SLIGHTLY transparent
`causalTypeColors.tsx`:
Added alpha value to all diagram element colors
(except Selected, which is just used for the border highlight. The border is staying solid intentionally)

This allows the user to see dependency arrows that were drawn UNDER elements, without significantly hindering
readability on elements themselves. The readability impact is definitly arguable though, so it would be good
to get this in a config value in the help menu later on.