React refactor: Toggle dependency, Select All
Implemented functionality for two more Element CRUD operations

==
Toggle Dependency:
I'm offering two different versions of this.

For the first, dependencies are toggled along a chain created from the selection
buffer, in the order elements were selected. This is the "Dependency Chain" button.
Example:
Selection Buffer: [1, 2, 3, 4]
Dependencies created: (1 -> 2), (2 -> 3), (3 -> 4)

For the second, dependencies are toggled along the order selected, where the target
of the dependency is always the last element in the selection buffer. This is the
"Dependency Group" button.
Example:
Selection Buffer: [1, 2, 3, 4]
Dependencies created: (1 -> 4), (2 -> 4), (3 -> 4)

Both versions prefer to fill in missing dependencies instead of removing any.
They will only remove dependencies when the selection already completely satisfies
the chain or group.

To make this easier, I adjusted elementAssociatedDependenciesMap, adding more info about the dependency
via a custom type and enum, defined at the top of CausalDecisionDiagram.
This incurred a couple minor changes elsewhere to use the new type instead of assuming it's just a UUID
string like before.

==
Select All:
Sets the selection buffer to the list of keys from the diagramElementsMap. This is done
this way because ElementCRUDPanel has easy access to this map already, so it's one line.
From testing, this seems to happen to select the elements in the same order they appear in
the original JSON list. Not sure if that is guaranteed by order of key insertion?