More sophisticated dependency selection styling
Dependency color and size styling now uses `getDependencyStyle`,
which sets these properties based on the dependency JSON and its
place in the selectionBuffer.

When only one element is selected, all dependencies connected to
the selected element are highlighted and enlarged.

When multiple elements are selected, the dependencies that are
connected to a selected element at both source AND target are
highlighted and enlarged. Elements connected only at one end
(source OR target) are semi-highlighted, and normal sized.

Added SemiHighlighted color to causalTypeColors for this.
For consistency, renamed Selected color to Highlighted. I didn't
want "SemiSelected".

Also moved DependencyRole and AssociatedDependencyData definitions
to the same file as causalTypeColors, now renamed to cddTypes.tsx.
There's probably a better place or name for these, but this should
at least be better than before.