React refactor: Element CRUD finished (for now)
Finished implemented the initial set of supported "Element CRUD" operations.

Implemented "Add Display to Element" function, with the option to select from all currently supported display types.
No display removal, but there's always undo and further edits can be done in the JSON editor.

To accomplish Add Display to Element, I modified all Display Type definitions to include functions for generating
the default JSON representation for each display type. Future Display Types have a clear pattern for implementation,
requiring a React Component definition, a Default JSON funciton implementation, and registration on the DisplayTypeRegistry.
With those accomplished, the display should automatically be supported in the display type options for the Add Display to Element button.

The above implied an architecture that Diagram Element and Dependencies did not follow. I added a Default JSON generation function
to DiagramElement.tsx for consistency, and created a new file in `lib` called `dependencyUtil.tsx` to hold the dependency
default JSON for now, until I come up with a better place for it.

Also:
- Improved readability for addNewElement by extracting new element placement logic to its own function
- Fixed all my CRUD operations that weren't ACTUALLY pure because they modified the state of selectionBuffer. Now
these functions either don't do anything to selection buffer, in which case the ElementCRUDPanel handles that side
effect, or they explicitly return the next state for SelectionBuffer as well, in which case ElementCRUDPanel still
handles actually APPLYING the state change.
- Fixed siutation where the JSON editor's expanded paths wouldn't update when dependencies were toggled, by setting
selectionBuffer to a copy of itself, updating the reference and triggering the paths useEffect in CausalDecisionModel