Fix: Added displays not expanded in JSON view
ALL relevant buttons now set the selection buffer in some way.
If no selections have changed, they set the buffer to a copy.
This registers as a change for React, since the reference is different,
which triggers expanded JSON paths to recalculate.

Previously, the "add display to element" button was not setting the buffer,
so new displays were not expanded in the JSON view.
"Delete display from element" now also sets the buffer, though I'm not sure
it's necessary. Consistency, though.