Fix stale selectionBuffer state when JSON chagnes
Fixed an issue where selectionBuffer should have updated as a side effect to JSON
changes (e.g. from the editor), but didn't.

This meant that you could
- Select an element in the graphical editor
- Go to the JSON editor, select all --> paste a completely different JSON object
that is missing the selected element
- Element stays selected, so you can select a new element in the new JSON
and attempt to toggle a dependency between the existing element and the now-nonexistant one

This is fixed now. selectionBuffer now verifies that all selected elements still
exist whenever JSON changes. This means selections don't break for EVERY change
to JSON (like position updates etc), but when the element is no longer present it can't
stay selected.