Dependency behavior: "Group" is now "Combine"
Slight terminology change after showing the new Element CRUD stuff to Dr. Roberts.
In UI, updated the word for this behavior to "Combine"
In code, updated some of the old references to "group" that made more sense as "combine".
Some still made more sense as "group" (in comments), but I clarified these a bit as well.