React refactor: Non-functional API settings menu
Added a "File" tab to the help/etc menu

This will hold model-level CRUD options (new/load/save/delete)
I'll be adding a "Save as New" button to assign a new UUID to an existing model.
I'll also have a *dedicated* "Download JSON" button.

This menu will also have the API option to change the base URL of your API.
If you update this URL, it will change where the app populates its options for Load/Delete,
and where it tries to save models to.

Also made some misc. changes to the existing menu and menu options.
- Moved JSON Editor to be the 2nd choice after File, taking the traditional place of an "Edit" menu
- Made the tab options vertically narrower. Added a top and bottom border to the selected option.
- Made the expand/collapse tab fit the size of the tab options better, in both landscape and portrait.
- Made the normal background color of menu sections darker, somewhat less blue
- Header styling is now a lighter blue stripe instead of a darker blacker stripe
- Adjusted spacing for inner menu content
- Moved some <ul> and <ol> objects outside of <p> tags, because my browser was complaining about that being invalid.

Next step: Implement all the functionality!
Also update the Help menu for the updated controls and contexts and whatnot