React refactor: API interface, APILess features
Defined an interface for API classes. Once the Go API is updated,
will be able to test a "real" implementation of the API interface.

For now, there is a "NoAPI" class that falls back on locally-available
functionality whenever API functionality is requested.

Remaining "File Tab" buttons to implement:
- New Model
- Update API URL (need "real" API class for this)

Dynamic API switching uses an APIHandler class that initializes
a new API whenever the base URL is changed.

Other changes:
- Fixed missing style rule on input boxes in the help menu
- Fixed many built-in models sharing the same model UUID
- Moved all built-in models to a designated folder
- Made a simple Record for retrieving all built-in models, in `src/model_json/builtin-examples/builtinModels.tsx`