Right menu cleanup/fixes
Fixed: Toggling the right-side menu is now a style-only change, using the `hidden` CSS class.
This means that state like the API base URL and selected model won't be lost when the menu is closed.
Also means the JSON editor no longer needs to be recreated each time the menu is re-opened.

Cleanup:
- Updated descriptions and instructions in the Help menu, defining all the new editor buttons
and settings from the File menu
- Modified style so that the menu tabs banner is a slightly different color from the default
background color in the menu tabs themselves. This way the banner sticks out a little more
when scrolled partway down e.g. the Help menu