Fix: JSON Editor doesn't scroll to selections
Fixes #68

See that issue's discussion thread for my analysis and conclusions.

Fixed an issue where the JSON editor was no longer scrolling to any content selections.
This had the knock-on effect of hiding the editor's top ribbon and bottom info displays if the user was scrolled partway through the JSON content.
All of this is now fixed.

- Added a style class name for the `<div>` that holds the JSON editor
- This style class now limits the editor's height to the height of its parent `<div>` and uses an `auto` overflow
rule for the JSON content, which is what the JSON editor needs in order to be able to control its own scroll position