Progress: Runnable editor views finished
(WIP)

Finished a first pass of the display component framework for the Runnable Model Editor screens.
Updated controls panel to list Controls that are used in the currently-open diagram separately from all other controls.
Updated controls panel to properly list displays for each control. The plan is to have a list of Displays that can be added to each
control, as a dropdown menu. This menu will list displays in the open diagram first, then displays for all other diagrams separately.

Implemented a selection buffer for IO values, which is shared between all components that list IO values.
This means that selecting an IO value on any panel selects it on EVERY panel, allowing you to more easily trace the flow
of data through the runnable model.