Fix minor display issues with runnable editor
Fixed: Runnable editor view throws off the horizontal space ratio between the left and right portions of the overall editor.
For sufficiently small browser viewports, this meant that the right menu would overflow out of the overall editor and hang off
the edge of the browser window. This also meant that switching between a diagram view and the runnable editor view would
abruptly change the size of the right side help menu.

For portrait viewports, fixed a similar issue where sometimes the help menu would cover the entire overall editor view.
This should no longer happen.

Also fixed an issue where
- Loading into the page in the default view
- Opening the help menu, going to File tab and editing the active runnable model
- Closing and then re-opening the help menu, and finally
- Opening the previously-opened diagram
would result in a bad state for XArrows.

I fixed this by swapping from
- using CSS to hide the diagram view when in the runnable editor view, and vice versa
to
- using a boolean flag to optionally add the JSX for the diagram view or the runnable editor *only*,
based on what is selected.
Basically, when the diagram view is not active, it is not part of the page's HTML. Same for the
runnable view.
This means that when the user switches back to the diagram view, it re-initializes. So it re-fetches
initial I/O values *and* re-draws XArrows.