Dynamic Ranges. Dynamic Selector input validation.
Ranges can now accept control parameters from I/O values, similar to dynamic selectors.
I/O values can be connected to range min, max, and step values. This way, ranges used "later" in a runnable
model can have their input space adjusted based on the current state of the rest of the model.

Added some minimal input validation for dynamic selectors. Now, when dynamic options
are provided, the currently selected option is validated against the list of available options.
If the selected option is not in the list of available options, it is automatically reset to the first
valid available option.

==
NOTE: This sort of validation is currently NOT being done on ranges in any context.
Ranges are sometimes used as output displays, and in other odd ways, so there's a little more
room for nuance on whether it's a good idea to do any or all of these things in any
arbitrary context:
- Ensure current range value >= min range value
- Ensure current range value <= max range value
- Ensure current range value is equal to min range value plus some integer multiple of the step value (though
this may or may not be true of the max range value itself, so that could be considered a soft fourth
validation possibility)

All of this behavior could be configured by flags in an Addon object.. But currently all Addon uses are
purely aesthetic/style settings, so this would set a different sort of precedent. For now, I'm just
leaving this alone.

Ultimately, Ranges can only provide Number values, but Selectors allow _any_ type for their option list.
It's easier to validate a potentially out-of-bounds range selection than it is to validate option types
and existance and whatever else. So for now, Selectors will just do that automatically.
==

Added Range Demo: Dynamic Range to demonstrate the new dynamic range behavior.
Adjusted Dynamic Selector Demo's summary to remove now-obsolete description about how the inputs weren't validated
when the options list changed.. because now they are!