New Eval Asset: API Call (basic, synchronous)
Added support for the API Call Evaluatable Asset defined here: https://github.com/opendi-org/cdm-json-schema/commit/1023d3f3e206d032c97510b0dbcdb468a930a62f

This is a minimal, basic version of API Call support, with lots of room for improvement regarding the way the model evaluation
processes handle API calls and other potential future *asynchronous* elements.

For now, API calls are naturally `await`ed so that future evaluatble elements that depend on their outputs don't break.
This means they're working properly, and they can be used in a way that is roughly equivalent to scripts, but it also means
that evaluation hangs on any API call until it receives a response.

For API calls to be most useful, a more async-minded approach will be necessary, but for now, that is a future TODO.

`evaluateModel.tsx`:
- Extracted functionality for evaluating individual Script elements into a function : `evaluateScriptElement`
- Added a new `evaluateAPICallElement` function for elements that use APICall Eval Assets
- Updated main evaluation loop. Now uses a couple switch cases to determine which evaluation function to call based on
the type of the eval asset the current element uses. Also switched to `for .. of ..` instead of `.forEach`, so I can `await`
the API Call evaluation
- Defined an `EvaluationResults` type, to standardize returns for both `evaluate..Element` functions
- `evaluateAPICallElement` is `async`, so `evaluateModel` is now also `async`
- `evaluateModel` now takes a map of Eval Asset data, so `evaluateAPICallElement` can more easily retrieve API call info from
the referenced asset's JSON data

`CausalDecisionDiagram.tsx`:
This is where `evaluateModel` is used, so I had to update it so that it can `await` the `evaluateModel` function.
- `computedIOValues` is now a `useState` instead of a `useMemo`. It is populated in a `useEffect` that is roughly
equivalent to the old `useMemo`, but with a React cleanup function that prevents state conflicts if the `useEffect` is
re-triggered while still `awaiting` the results of `evaluateModel`
- New `evalAssetMap` memo, populated with the new model preprocessing function, for passing to `evaluateModel` for API call info retrieval

`modelPreprocessing.tsx`:
- `getFunctionMap` now explicitly returns a map with `Function` values, instead of `any`
- Slightly refactored `getEvaluatableAssetFunctionNamesMap` to rely on other existing functions more
- `getFunctionNamesFromAssetMap` ensures the Eval Asset is a Script before trying to read functions from it

`RunnableModel.tsx`:
- Adjusted the display for individual Evaluatable Elements to account for the existance of API Call assets. Now,
the element only gives options for script names if the element references a Script. If it references an API call,
displays relevant info about that API call instead

`RunnableModelEditor.tsx`:
- Adjusted the display for idnividual Evaluatable Assets to account for the existance of API Call assets. Now,
API Call assets don't have an "Edit APICall" button (use "reveal in JSON to edit JSON directly instead"), and
they'll display relevant info about the API call URI and method.
- Added an "Add New API Call" button, equivalent to the "Add New Script" button. Calls `addAPICallToModel` from `runnableCRUD.tsx`

`runnableCRUD`.tsx:
- Added `addAPICallToModel`, which adds the default API Call JSON (defined in `defaultJSON.tsx`) to the model's evaluatableAssets list

`defaultJSON.tsx`:
- Added `defaultAPICallJSON()` which returns default JSON for an API Call typed Evaluatable Asset. This currently
fills in a basic example API call from https://jsonplaceholder.typicode.com/, with a GET method, no payload, and now URI extension,
for its default asset content

`ControlText.tsx`:
- Minor bugfix: Text control now always casts the value it is intended to display to a String before using it as such