Minor fixes for local debugging QOL
Submodules URL now uses HTTP instead of SSH.
This means that it should be easier to clone the submodules, but harder to push to them from within the submodules here.
This is fine, since it's better to edit the submodule from its own repo anyway.

Updated Vite config file to remove the trailing slash requirement from the base URL definition.
This means that locally you can use either http://localhost:80/cdd-authoring-tool or http://localhost:80/cdd-authoring-tool/ (with the trailing slash) to view the page.

Updated README to be more accurate.

Still todo: Test workflows on other operating systems and further update README if necessary.