Footer: Best-guess links for blog and glossary
Updated placeholder links for Blog and Glossary in the footer.
Rather than just linking to the top level opendi.org, Blog now links to opendi.org/blog and Glossary links to opendi.org/glossary.
I think this is what the URLs will be once the Gamma site goes live, but if not we'll just need to update them again later.